Address feedback 2/2
A python wrapper is implemented in order to pass the
`-datadir=$XDG_DATA_HOME` flag. This allows to remove the
`--persistent` flag.

The desktop file was copied from
[here](https://github.com/bitcoin-core/packaging/blob/master/debian/bitcoin-qt.desktop),
with the only changes being the icon and Exec fields. It would be better to
download directly from the url and apply a patch. An even better solution would
be to add the desktop file, together with the appdata and
icons to the distributed tarball.

`test_bitcoin` is failing on 32bit archs, `flathub.json` was
re-enabled in order to only build for 64bit.